Better structuring of rawError and profileLikelihood
These now are set aside as internal functions (just what I'm calling them, nothing intrinsic to them) and options are passed into them as structs and a call to overrideDefaultOptions uses this info. User-facing functions that end up calling them accept name-value parameters that can set their option values. Intermediate functions gather these into structs to pass in. So, for the most part, rawError and profileLikelihood are called with the full set of options passed in. The one exception is that I allow for rawError in testSMFitToABM.m to load a raw_error_opts struct or initialize an empty one. The obvious solution is to make sure those files save the entire raw_error_opts struct.