Sensitivity in KerrisABM and TrivialModel
Got KerrisABM set up with the new sensitivity workflow. Also, made TrivialModel to test what global sensitivity says in a very simple and controlled context. I learned that GS depends on both the model and the distribution on parameter space. Having a uniform distribution on a parameter makes the GS outcome very dependent on the distribution parameters (min and max). Truncated normal is much less sensitive to the min and max values, though it will of course also be sensitive to the mean and std.

What is amazing (and appears to be true across all these models) is that our indirect GS method for the ABM/CM (complex model) accurately reproduces the GS based on the direct method.