Indirect GS workflow allows for using best SM
Now can pass an optional parameter to indicate using the optimal parameters for GS rather than the profiles.

Also, better use of files where I'm going to start saving all previous file names in new files. So the profile file will also store a struct "files" that points to the data file and the optimal parameters file.

This is only implemented for the TrivialModel.

Also, sampleFromSM is split into sampleFromSMProfiles and sampleFromSMBest for the two ways to "sample" from the SM (is best really "sampling"?). These are no longer user-facing. Instead, setupSampleFromSMFunction returns a handle to the appropriate one of these two.